Move Celery Rollbar stuff to celery_rollbar.py.
